Create suggestion-list at CitySearch.js
Import useState. Create state variable, called "showSuggestions" with initial state "false". Use onFocus event handler of the <input /> element, so if the the input field gained/lost focus change the state of "showSuggestions".  Use ternary operator to render a <ul> element with a suggestions list, if "showSuggestions" is true. Expect tests  "suggestions list is hidden by default" & "renders a list of suggestions when city textbox gains focus"  to be passed.