Change suggestions-list depending on user-input.
Update CitySearch.js file to change the suggestions list depending on the change of input field. Create state variables, called "query" and "suggestions". Use the "query" state as the value of the input field. Create "handleInputChanged" function and set it as the callback function of "onChange" on the <input/> element. Return items from the list of "suggestions" and use "suggestion" as a key to distinguish each item.  Expect "updates list of suggestions correctly when user types in city textbox" test to be passed.