Create "fetchData" function at App.js
Import "useEffect","useState" and "getEvents". Create state variables, called "events" and "currentNOE". Pass the events state as a prop to EventList. Create fetchData function to populate the "events" state with the events list, that will be fetched. Use useEffect to call fetchData. Expect test "renders a list of 32 events when the app is mounted and rendered" to be passed.