Update App.js&CitySearch.js for integration test
Import extractLocations function, create state allLocations variable, initialize it in fetchData() function and pass it as a prop inside of App.js file. Import useEffect and use the stringified value of the allLocation prop as a dependency in useEffect at CitySearch.js file. Expect "renders suggestions list when the app is rendered" integration test to be passed and "renders a list of suggestions when city textbox gains focus" unit test to be failed.