Create getAccessToken() at api.js
Create and export a new async function called getAccessToken. Try to find an access token in the localStorage.   Check for an authorization code, if no token is found (!accessToken).  Redirect user to the Google Authorization screen, where they can sign in and receive their authorization code, if no authorization code is found (!code).